refactor: move class to different package for better organization
